+++ packages/suite-base/src/panels/diagnostics/DiagnosticStatusPanel.style.ts
@@ -1,0 +1,9 @@
+// SPDX-FileCopyrightText: Copyright (C) 2023-2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)<lichtblick@bmwgroup.com>
+// SPDX-License-Identifier: MPL-2.0
+import { makeStyles } from "tss-react/mui";
+
+export const useStyles = makeStyles()({
+  toolbar: {
+    paddingBlock: 0,
+  },
+});
